move actual analysis to own hook
